Detect and correct notes which are overlapping
Thanks to gloomy for the heads up on this. If two notes in the same track start and finish on the same tick , the N64 will glitch out an instrument and cause it to incorrectly play.
This occurs naturally when trying to convert the Weldar boss theme from Banjo-Tooie, but with this simple script can now be automatically corrected upon conversion